Fix Atom losing focus breaking the focus terminal command
If the user switched away from Atom while the terminal had focus,
subsequent uses of the focus terminal command would block most keybinds.
This was happpening due to the terminal setting itself to the last
focused element when Atom regained focus. Refusing to set terminal
elements as the lastActiveElement solves the issue.